% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_animalcules_MAE.R
\name{clean_animalcules_MAE}
\alias{clean_animalcules_MAE}
\title{Clean up taxon names in a MultiAssayExperiment}
\usage{
clean_animalcules_MAE(dat)
}
\arguments{
\item{dat}{A \code{MultiAssayExperiment} object specially formatted as an
animalcules output.}
}
\value{
An animalcules-formatted \code{MultiAssayExperiment} object with
  cleaned-up taxonomy nomenclature.
}
\description{
This functional is an optional method for fixing up taxon names in a
\code{MultiAssayExperiment} to be run before \code{filter_animalcules_MAE}.
Specifically it removes brackets from species names, replaces species labeled
as "others" with "sp." and finally replaces underscores with spaces.
}
\examples{
in_dat <- system.file("extdata/MAE_small.RDS", package = "LegATo") |> readRDS()
clean_animalcules_MAE(in_dat)

}
