% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_hotelling_t2.R
\name{.Hotelling_mv_T2}
\alias{.Hotelling_mv_T2}
\title{Paired Multivariate Hotelling's T-Squared Test}
\usage{
.Hotelling_mv_T2(input, Group1, Group2)
}
\description{
This function performs a paired version of a multivariate
Hotelling's T-squared test.
}
\details{
Null hypothesis: population mean vectors are equal
Alternative: mean vectors are not equal

Source: https://online.stat.psu.edu/stat505/lesson/7/7.1/7.1.4
Example: https://online.stat.psu.edu/stat505/lesson/7/7.1/7.1.9
}
