% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stacked_data.R
\name{get_stacked_data}
\alias{get_stacked_data}
\title{Documentation}
\usage{
get_stacked_data(dat, taxon_level = "genus", covariate_1, covariate_time)
}
\description{
Documentation
}
\examples{
in_dat <- system.file("extdata/MAE_small.RDS", package = "LegATo") \%>\% readRDS()
get_stacked_data(in_dat, "genus", covariate_1 = "Sex", covariate_time = "Month")

}
