% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_hotelling_t2.R
\name{test_hotelling_t2}
\alias{test_hotelling_t2}
\title{run_Hotelling_mv_T2}
\usage{
test_hotelling_t2(
  dat,
  test_index,
  test_time,
  taxon_level = "genus",
  num_taxa,
  grouping_var,
  paired = FALSE,
  pairing_var = NULL,
  unique_id_var = NULL
)
}
\description{
run_Hotelling_mv_T2
}
