% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LegATo-package.R
\docType{package}
\name{LegATo-package}
\alias{LegATo}
\alias{LegATo-package}
\title{LegATo: LegATo: Longitudinal mEtaGenomic Analysis Toolkit}
\description{
LegATo is a suite of open-source software tools for streamlining longitudinal microbiome analysis, integrating visualization, modeling and testing procedures.
}
\author{
\strong{Maintainer}: Aubrey Odom \email{aodom@bu.edu} (\href{https://orcid.org/0000-0001-7113-7598}{ORCID})

Other contributors:
\itemize{
  \item Yilong Zhang \email{elong0527@gmail.com} (Author of NMIT functions) [contributor]
  \item Jared Pincus \email{pincus@bu.edu} (\href{https://orcid.org/0000-0001-6708-5262}{ORCID}) (Naming consultant of package) [consultant]
  \item Jordan Pincus \email{pincus@bu.edu} (Artist of LegATo icon) [artist]
}

}
\keyword{internal}
