% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MAE_SE.R
\name{parse_MAE_SE}
\alias{parse_MAE_SE}
\title{Parse a MultiExperimentAssay object and extract the elements as data.frames}
\usage{
parse_MAE_SE(dat, which_assay = NULL, type = "MAE")
}
\arguments{
\item{dat}{Either a \code{MultiAssayExperiment} or a
\code{SummarizedExperiment} object.}

\item{which_assay}{Character string. This refers to the assay to be extracted
from the MultiAssayExperiment object if \code{type = "MAE"}. Does not need
to be specified if \code{type = "SE"}. Default is \code{NULL}.}

\item{type}{One of "MAE" denoting a \code{MultiAssayExperiment} or "SE"
denoting a \code{SummarizedExperiment}.}
}
\value{
Returns a list of 3 named data.frame elements, \code{counts}, \code{sam}, and
\code{tax} denoting the counts data, sample metadata table, and taxonomy table,
respectively.
}
\description{
This function takes an animalcules-formatted \code{MultiAssayExperiment}
object and parses it to extract a named assay alongside the taxonomy and
metadata.
}
\examples{
# example code

}
