% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{run_gee_model}
\alias{run_gee_model}
\title{Compute generalized estimating equations (gee)}
\usage{
run_gee_model(
  dat,
  taxon_level = "genus",
  unit_var,
  fixed_cov,
  corstr = "ar1",
  plot_out = FALSE,
  plotsave_loc = ".",
  plot_terms = NULL,
  ...
)
}
\description{
Run an independent GEE model for each taxa with relative abundance
Works well with small data - multiple subpoints/subjects across clusters
}
\details{
https://data.library.virginia.edu/getting-started-with-generalized-estimating-equations/

fixed_cov is a vector
}
